[refactor] header를 common components 에서 module로 변경
