+++ vite.config.ts
@@ -11,6 +11,7 @@ export default defineConfig({
       '@common': fileURLToPath(new URL('./src/common', import.meta.url)),
       '@core': fileURLToPath(new URL('./src/core', import.meta.url)),
       '@modules': fileURLToPath(new URL('./src/modules', import.meta.url)),
+      '@^header': fileURLToPath(new URL('./src/modules/header', import.meta.url)),
       '@^login': fileURLToPath(new URL('./src/modules/login', import.meta.url)),
       '@^stuffs': fileURLToPath(new URL('./src/modules/stuffs', import.meta.url)),
       '@^histories': fileURLToPath(new URL('./src/modules/histories', import.meta.url)),
